+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in
 this file.
 
 ## UNRELEASED
+- Added tape compression processor.
 - Improved hysteresis engine performance by ~1.5x.
 - Improved oversampling menu with choices for linear vs. minimum phase and real-time vs. offline rendering.
 - Made oversampling choices preset-agnostic.